Correcting the issue where an expression which we are runtime checking the value of would be duplicated, also duplicating any side effects it may have had.
We achieved this by saving it into a runtime variable, and referring back to
that variable.  The variable is named with a $ which is invalid in java but
valid in bytecode so as to avoid name collisions with java programs.  Name
collisions with java compiler plugins of this type are still possible.

Attribution is accomplished in a separate tree translation phase, statements to
be attributed are collected, and then after all tree modifications are done the
attribution is completed.  This is to ensure we don't try to attribute any trees
which are not finished all edits, which could result in an invalid tree.

Tree translators have been refactored and enhanced, so that we have a
GeneralTreeTranslator to provide a "visitTree" capability.  Then a runtime check
tree translator to insert runtime checks, and a single edit tree translator for
small changes like replacing value references with variable references.  Finally
there is an attributing tree translator which will attribute a list of
statements in the tree.