Fixed issue where variable declarations which were the statement being runtime checked would end up in a different scope.
Solve this handling variable declarations differently, we first create a new
declaration, with no initializer, then create an assignment to that variable,
with the initializer as the value, and treat this as the statement being
checked.  We then edit the JCBlock that this is applying to and insert the
statements into the block at the proper location.

Also added a test to validate this behaviour.